simplify template tests
